+++ src/Application/Deployment/ProductionAutoloadGenerator.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace EventFlow\Application\Deployment;
+
+interface ProductionAutoloadGenerator
+{
+    public function generate(string $packageDirectory): void;
+}